EEPROM Config Sub Menu + other updates (#183)
* Fix compile errors in WavTrigger

These changes mirror the proposed changes to the original WavTrigger library which resolve several compiler error messages when compiling code for the Neutrona Wand or Proton Pack.

* Fixes, binaries

This fixes a typo in ATTENUATOR.md which was triggering check fails, fixes a lack of enclosing scope in the CheckSerial1() function in ProtonPack.ino, fixed an unhandled case statement in the Attenuator's Bargraph.h, and added precompiled binaries for the changes present in this PR.

* Add custom binaries, rename folder

Renamed version folder to 2.0.1, added updated custom binaries for FruttoVentLight and QuickVent.

* Delete erroneous file

Delete accidentally included binary from source folder.

* Whitespace updates

Updates whitespaces in several files, and removes now-obsolete specific binaries as well as a leftover file in the Neutrona Wand folder

* Correct Wand power levels

Changed wand power levels to use correct sounds for each level.

* Revert "Correct Wand power levels"

This reverts commit b22f989d14558a08727d779e4ce6dd75b1457e20.

* Neutrona Wand Fixes

Corrects enum typo in function to call bootup sound, corrects bug in 1984/1989 mode where turning off the Activate switch while the Vent/Lower switch is on would cause the shutdown sound to play twice overlapped.

* Fix CTS Mix Check

Corrected inverted boolean on a CTS Mix firing check.

* Whitespace reversion

Reverts github adding unnecessary whitespace to EOF

* EOF Whitespace fix

Second time's the charm?

* Whitespace adjustment

* Fix SFX overlap in CTS Mix

This fixes overlapping Proton Stream sound effects when going in and out of CTS Mix firing when the Neutrona Wand is in Power Level 1 through 4.

* wip wand updates

* config eeprom sub menu

* Add startup sound for standalone wand

This adds a startup sound to the wand if it is in Afterlife mode and is not being used with a GPstar-equipped Proton Pack. This allows the wand to sound a little more "complete" when first turned on since there is no pack connected to make those sounds instead.

Also fixes a few minor typos.

* Update CTS Mix Functionality

This proposed change makes CTS Mix into an independent optional fire mode that can be used alongside the video game or spectral fire modes. In this version, CTS Mix works only when the user is firing a Proton Stream and then holds down the Barrel Wing Button. If they release it, it returns to Proton Stream as before. If they release Intensify, the wand stops firing. When not firing, the Barrel Wing Button continues to be a mode switch as if CTS were not enabled.

* removed unwantedd wt comments

* cts updates, prelim menu system for eeprom sub level

* additional sounds for new sublevel eeprom config menu

* additional sounds

* eeprom config sub menu system

* eeprom config sub menu updates

* compiler checks to shrink code for diy builds

* invert bargraph for all options

* updated binaries

* documentation update

* Minor text fix, correct variable name

Assumed the "vga" in "b_vga_mode" was intended to be "b_vg_mode". Corrected new code before it became standardized.

Made a few spelling/grammar fixes, long with one

* Typo on Operations

* Caught by spellcheck

---------

Co-authored-by: nomakewan <nomakewan@users.noreply.github.com>
Co-authored-by: Dustin Grau <115430732+DustinGrau@users.noreply.github.com>